refactor(baseball): restructure classes with new packages
