+++ src/main/java/baseball/Application.java
@@ -1,5 +1,7 @@
 package baseball;
 
+import baseball.staff.StageManager;
+
 public class Application {
     public static void main(String[] args) {
         var stageManager = new StageManager();
