+++ src/main/java/baseball/giveup/GiveUpException.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.giveup;
 
 public class GiveUpException extends Exception {
     public GiveUpException(String message) {
