+++ src/main/java/baseball/playground/GameRecorder.java
@@ -1,5 +1,6 @@
-package baseball;
+package baseball.playground;
 
+import baseball.valueholder.PlayerAnswer;
 import java.util.ArrayList;
 import java.util.List;
