+++ src/main/java/baseball/staff/GameFinalizer.java
@@ -1,4 +1,8 @@
-package baseball;
+package baseball.staff;
+
+import baseball.setting.Buttons;
+import baseball.setting.Texts;
+import baseball.util.ValidatingUtils;
 
 public class GameFinalizer extends ConsoleInputForm {
     public boolean isRepeating() {
