+++ src/main/java/baseball/util/ConvertingUtils.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.util;
 
 import java.util.Arrays;
 import java.util.HashSet;
