+++ src/main/java/baseball/valueholder/CorrectAnswer.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.valueholder;
 
 public record CorrectAnswer(String answer) {
 }
