+++ src/main/java/baseball/valueholder/InitialSettings.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.valueholder;
 
 public record InitialSettings(String gameType, int gameSize) {
 }
