+++ src/main/java/baseball/valueholder/PitchingResult.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.valueholder;
 
 public record PitchingResult(int ball, int strike) {
 }
