+++ src/main/java/baseball/valueholder/PlayerAnswer.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.valueholder;
 
 public record PlayerAnswer(String answer) {
 }
