+++ src/test/java/baseball/UmpireTest.java
@@ -2,6 +2,9 @@ package baseball;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import baseball.playground.Umpire;
+import baseball.valueholder.CorrectAnswer;
+import baseball.valueholder.PlayerAnswer;
 import org.junit.jupiter.api.Test;
 
 class UmpireTest {
